Moving language-agnostic components into a separate subproject
