+++ language-server-commons/src/main/java/com/palantir/ls/server/util/SourceWriter.java
